#15: move all lessons; update commands; create, expand test files
