+++ package.json
@@ -1,19 +1,19 @@
 {
   "name": "intro-to-code-js",
-  "version": "2.1.0",
+  "version": "2.2.0",
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test-01": "mocha './01-Values-and-Data-Types/**.js'",
-    "test-02": "mocha './02-Conditionals/**.js'",
-    "test-03": "mocha './03-Methods-and-Functions/**.js'",
-    "test-04": "mocha './04-Arrays-and-Loops/**.js'",
-    "test-05": "mocha './05-Callbacks-Iterators/tests/**.js'",
-    "test-06": "mocha './06-Objects/**.js'",
-    "test-07": "mocha './07-Objects-2/tests/**.js'",
-    "test-08": "mocha './08-Recursion/tests/**.js'",
-    "test-09": "mocha './09-Classes/tests/**.js'",
-    "test-10": "mocha './10-Data-Structures/tests/**.js'",
+    "test-01": "mocha './lessons/01-Values-and-Data-Types/tests/**.js'",
+    "test-02": "mocha './lessons/02-Conditionals/tests/**.js'",
+    "test-03": "mocha './lessons/03-Methods-and-Functions/**.js'",
+    "test-04": "mocha './lessons/04-Arrays-and-Loops/**.js'",
+    "test-05": "mocha './lessons/05-Callbacks-Iterators/tests/**.js'",
+    "test-06": "mocha './lessons/06-Objects/**.js'",
+    "test-07": "mocha './lessons/07-Objects-2/tests/**.js'",
+    "test-08": "mocha './lessons/08-Recursion/tests/**.js'",
+    "test-09": "mocha './lessons/09-Classes/tests/**.js'",
+    "test-10": "mocha './lessons/10-Data-Structures/tests/**.js'",
     "test-twitter": "mocha './11-Projects/Twitter/tests/**.js'",
     "test-debug": "mocha './debug/tests/**.test.js'",
     "test:mochawesome": "mocha './**/**.js' --reporter=mochawesome --reporter-options autoOpen=true",
